Refactor: api core 함수 변경에 따라 전체 댓글 또는 한 개의 댓글 GET 요청 시 action 함수의 리턴 값 변경 +파일 이름명 및 삭제에 따른 경로 수정
